refactor benchmarks to use number metrics
